Support shorter tokens used with authentication (#3)
This adds a new send_links configuration option which defines whether to send links for the user to click in renewal emails, and defaults to true. If set to false, only the renewal token is sent, which is expected to be copied by the user into a compatible client, which would then send it to the homeserver in an authenticated request (which means we don't need to ensure these shorter tokens are unique across all of the users).